데이터 수정 추가 & match case 에서 정해진 값이 나오지 않으면 나는 에러 해결

+++ src/main/scala/api/phoneBook/Manager.scala
@@ -11,7 +11,19 @@ class Manager {
     phoneInfoRepository.saveData(phoneInfo);
   }
   
-  def deleteDate(name: String) {
-    phoneInfoRepository.deleteData(name)
+  def deleteDate(name: String): Boolean = {
+    if(phoneInfoRepository.isExistData(name)){
+      phoneInfoRepository.deleteData(name)
+      true
+    }
+    false
+  }
+  
+  def updateDate(phoneInfo: PhoneInfo): Boolean = {
+    if(phoneInfoRepository.isExistData(phoneInfo.name)){
+      phoneInfoRepository.saveData(phoneInfo);
+      true
+    }
+    false
   }
 }
